int 32

+++ bip0009_test.go
@@ -27,7 +27,7 @@ const (
 
 	// vbTopBits defines the bits to set in the version to signal that the
 	// version bits scheme is being used.
-	vbTopBits = 0x20000000
+	vbTopBits = int32(0x20000000)
 )
 
 // assertVersionBit gets the passed block hash from the given test harness and
